feat: cross compile for PI
Command for cross compilation:
`cargo build --release --target aarch64-unknown-linux-gnu --bin {}`
Then transfer ths binary to PI using `src target/... IP:~/`

The `openssl` dependency that is added here is only for the purposes of
cross compilation, this project doesn't directly depend on OpenSSL.

But since rocket requires it, the cross compilation target requires a
version for arm linux to be present on the system, by adding this
particular line with that specific feature the openssl is also compiled
from source during compilation and we do not get any errors. Otherwise
we had an error relating to the openssl not being found for arm linux
target on system.